Implement DrawDabs command
Add a new DrawDabsClassic command that draws soft brush dabs.
(Hard edged brush is not yet reimplemented.)
Update Freehand tool to use the refactored ClassicBrushState class
to generate dabs. (Works in soft and watercolor mode)

ToolChange and PenMove commands and related code is not yet removed,
but their handling is disabled in StateTracker class. Both PenMoves
and DrawDabs are generated by the Freehand tool so that recordings
can be loaded with a dev-2.1 branch version for comparison.

Other brush based tools have not yet been updated and do not work.