Add TrustedUsers command
Add a new command to the protocol for listing trusted users.
This works similarly to the operator status, except the
interpration of "trusted" is entirely clientside. The server
remembers the status for authenticated users, though.

Implement also the GUI for showing and toggling the status.
Filtering based on this new access tier is not yet implemented.