Keep just the minimum session history
Clientside: release all session history older than the oldest snapshot.
There is no need to keep anything more than that.
(Except for one thing: if the whole session history is in memory, a
recording can be started without making a new snapshot. However, this
is no longer possible once the beginning of the history is lost, after which
all the extra history is just useless baggage.)