Allow different input settings per brush
The input tab now allows specifying multiple different smoothing and
pressure mapping configurations. Every brush slot lets you pick which of
these configurations you want to use for it.

This lets you, for example, create a "sketch" configuration with little
smoothing assigned to brush 1 and a "line" configuration with lots of
smoothing for brush 2.

Resolves #915.