Convert the input dock into a dialog
Now that an input preset can be assigned to each brush slot,
it's no longer necessary to edit the input curve as often as before.
The input edit view can now be shown in a dialog box instead, which pops up
from a button next to the preset selection. (Should be easier to find this way too.)