Make root paths of the proxy server configurable
Instead of collapsing the WebSocket URL and the file serving into the
same root path, they now have two configuration options -files-root
(defaulting to /) and -ws-root (defaulting to /ws). They can be set to
different values to allow for a different root to serve from.

The default WebSocket URL of the React frontend also now has /ws
appended to match with these defaults and passes the configureed
PUBLIC_URL to the Emscripten createModule function.

And finally, default the PUBLIC_URL in the React frontend to `.` to make
arbitrary paths work.