Protect against missing extern "C"
By adding a check in dpcommon/common.h that ensures DP_INSIDE_EXTERN_C
is defined. Some editors like to sneakily add includes on autocomplete,
which causes linker errors when including the Drawdance C headers
without the appropriate extern definition. With this commit, the
compiler will error out and show where the offending include is located.
Non-malicious includes have to #define DP_INSIDE_EXTERN_C and #undef it
afterwards now, but since the actual inclusion locations are few and far
between, that's not much an overhead.