MyPaint support and rework of brush presets docker
In addition to supporting MyPaint brushes, it now behaves more like
Krita's presets docker in that it uses tags instead of folders. A single
brush can be put into multiple tags. Brushes can also be named, given a
description and a custom thumbnail now. Sort order is now dependent on
name, which is what MyPaint and Krita do, rather than having the user
push them around. Searching of brushes is also now supported, since they
got names to search them by. And of course you can import MyPaint
presets.

The internals are reworked to just use SQLite instead of dealing with
bespoke files and messing around with that timer. Old presets are
converted to the new format on the first run through.