Add forceNsfm config option for the server
To always force NSFM mode on all sessions, which is mandatory on several
servers. When the config option is set, attempts to unset NSFM from a
session are ignored. Setting this flag while the server is running will
automatically set the NSFM flag on all sessions.

The client part will follow in a separate commit, but clients don't
actually need to understand its existence, so this is split off for
easier cherry-picking.