Start replacing Rustpile with Drawdance
Switches out the paint engine from the Rust-based one to Drawdance. The
advantage being that it's faster and runs in the browser, meaning there
would only be one paint engine maintained.

This is a work in progress state, some features like playback of
recordings or animations don't work properly yet. However, this is still
a pretty stable state.

The Rust code has not yet been removed, since it's still useful for
reference.