Implement recording playback
Including reading and writing of index files. Also adds resetting of the
canvas history to a specific canvas state through an internal message,
which is required to skip through recording indexes.

It seems like sometimes index snapshots aren't written properly, but I
guess I'll fix that in a later commit. Most of them work okay, so it
hopefully shouldn't be too bad.