Add debug dump recording to developer tools
A checkbox with a warning message behind it that lets you turn on debug
dumps. They'll record to the dumps directory in Drawpile's data
directory (%APPDATA% or .local/share or equivalent) after the next
canvas reset if enabled. This is not remembered across restarts and
windows intentionally, since the dump files get kinda large.