Add Link, LinkVisited and PlaceholderText colors
So that those elements look right in the different themes. Links were a
really dark blue that was hard to see in dark themes and the placeholder
text was black, which really only showed up in Qt 6, but there it made
placeholders in text input fields really hard to read. Also updates the
Krita theme conversion script as well as all theme colors accordingly.