Replace dock buttons and menu with menu button
Since closing and undocking docks is done pretty rarely, I think it's
fine to turn it into a menu instead. Replaces the troublesome custom
button with a regular QToolButton that gets a new icon not used anywhere
else in the application. Also gives easier access to global dock options
such as locking docks or toggling move mode with shift, making that a
bit easier to find.