Allow setting minimum brush pressure values
Replacing the toggle buttons on the right of the brush dock with popup
windows with a checkbox to toggle pressure sensitivity and a slider to
change the minimum value. Instead of the minimum being a fixed value,
it's now a percentage, which is how other drawing programs handle it
too. Since you can now no longer tell if pressure is turned on from the
button, the prefix of the sliders now says the beginning of the range,
if applicable. The switch from the nondescript brush button to a
settings button should hopefully also cut down on people being unable to
figure out how to toggle pressure.