Use a DP_Mutex instead of a QMutex in paint engine
To lock access to the QPainter used for rendering on the canvas pixmap.
QMutexes don't just use pthread or Win32 mutexes, instead they implement
something custom. This has shown itself to be unbelievably slow with
semaphores in Win32, so I don't trust mutexes not having the same issue,
hence replacing this hot one. Drawdance mutexes are just thin wrappers,
so we know they are sane.