Specify Clang root explicitly for Ubuntu CI
Apparently CMake lookup is non-deterministic enough, or the CI
environment changes enough, that leaving it to the default CMake
logic results in a different clang version being picked sometimes
with missing files (because that version of the -dev package is
not installed) that breaks the build of Qt.

Trying to set env on the matrix does not appear to work, and trying
to set env on the step using a value from the matrix also does not
seem to work, so it echoes to $GITHUB_ENV, even though this is
clearly silly.