Wake and Wifi lock on Android when connected
Keeping the device awake and internet working while connected to a
drawing session so that it doesn't drop out, or at least not so easily.
I'm not totally sure if this actually works properly, documentation
suggests that it shouldn't, but from testing it seems to hold up fine
when the application is sent to the background or the screen is turned
off.