Add a shim for (Qt|QString)::SkipEmptyParts
Since that got changed between Qt versions. Replaces all ad-hoc handling
of this dichotomy with the use of that shim. Also rewords qtshims.h a
bit to separate each version-specific shim into its own section so that
it can account for more than just Qt 5 and 6, as well as enabling the
QtGui shims only when the current target is actually linked against
that.