Make file saving work somewhat sanely on Android
By asking the user for a name and type ahead of time and then hoping
they don't mess it up by deleting the extension in the native file
picker. If they do, we yell at them to not do that.

Also separates the single "last path" setting into several categories so
that they don't clobber each other, sets a proper file extension for the
last path and properly attaches a file extension to the chosen path
(except on Android, where we can only yell at the user to do it right.)