Expand the abilities of classic brushes
Giving them options for a minimum value for size, opacity, hardness and
smudging, as well as pressure curves for them, rather than them always
behaving linearly. Also stores the eraser state as well as the blend
mode inside the brush now, kind of like MyPaint brushes do, so that half
of that implementation isn't stuck in the frontend.