Change how indirect mode flag is saved
Instead of direct mode being in the "inc" key, now indirect mode is
saved in the "indirect" key. This is because the old, broken indirect
mode was basically just a bootleg direct mode, so it's likely that
people have been using it as such. Hence the behavior is closer to
direct mode.