Implement a user info dialog
Right-clicking on a user in the list next to the chat now has the option
of showing information about them. Doing so will display a bunch of
technical stuff that helps diagnose errors without having to ask them to
find all that information manually and writing it into the chat.