Handle clearing and cutting selections properly
Removes a jerking around of the preview by not temporarily altering the
shape an instant before the selection is cleared and behaves properly
when a clear is requested while a move is in progress, it caused the
area underneath the selection to be cleared while keeping the transform
going, which isn't sensible behavior. Now it just deletes what was
selected and gets rid of the selection.