Fix broken builds and missing install targets
This commit rewrites the existing CMake files so that they
export targets and install required headers in a way that allows
the application and libraries to install and be idiomatically
integrated with other projects.

Included third-party code now builds object libraries that are
embedded in the final archives and public headers from those
third-party libraries are installed in subdirectories of the
embedding libraries.

This commit also fixes compiling the drawdance and dpgles2
targets and the emscripten cross-compilation, which were all
broken.

Across CMake files, verbose copyright headers are replaced with
SPDX identifiers plus an AUTHORS.txt for standard attribution.

Finally, this commit removes clang-format from CMake, since this
cannot work with all generators according to
<https://discourse.cmake.org/t/clang-format-integration/3358/4>
and should go into a git pre-commit hook instead or integrated in
the IDE.