Add MyPaint brush permission
Since they can be abused somewhat easily, if a particularly terrible
brush is used to cause way too many dabs or something.

Introduces a new tool locked state and making it clear in the brush
settings docker that the user isn't allowed to use MyPaint brushes, to
avoid confusion as to why they can't draw anything.