Add classic brush stabilizer, remove input presets
Classic brushes now support a stabilizer like MyPaint brushes do, with
the stabilizer level stored in the brush itself. This is available as a
slider in the brush settings dock now.

The stabilizer stops when your cursor stops moving, it doesn't catch up
like it does in SAI or Krita. This is intentional, since the MyPaint
stabilizer works the same way. Could be fixed, but would require patches
to MyPaint. I guess we'll see if there's significant desire for it.

Input presets have been largely removed, since we don't need two ways to
control that kind of stuff. Pressure curves are available for every
parameter of every brush, so there's no need to have a global curve as
well. Velocity and Distance options were removed since MyPaint brushes
have way better, much more useful options in that regard. The averaging
stroke smoother is still there, since I can imagine that's at least
useful if you have a very jittery tablet, but it has been banished to
the preferences dialog, since you're not longer expected to fiddle with
it constantly, given that there's a stabilizer slider for that purpose
now.

Fixes #179.