Track laser and move cursor changes properly
Keeping only the latest information doesn't work properly, since the
frontend relies on getting a persistence of zero to stop a laser trail,
which might be missed if we clobber it. So now it is tracked as a vector
of events. Also splits the crazy re-used meta buffer into separate ones,
we're not in need of saving those few bytes.