Fix missing dependencies in Windows builds
It looked like windeployqt took care of bundling all dynamically
linked libraries so long as the DLLs were in its prefix directory,
but it turns out this was a lie. It only appeared to ever work
because the only thing that had been dynamically linked before
Dancepile was added to the project was QtKeychain, which happens
to start with `qt`, which made windeployqt think it was part of Qt.

This uses the same trashy dependency walking that is used for
Android because CMake still has not figured out how to do this
itself in a way that is sane.