Allow indirect mode on MyPaint brushes
Because it was requested and we got a bit left over in the protocol for
it anyway. MyPaint doesn't have an indirect mode, so brushes aren't
intended to be used this way. Some features get disabled, such as
smudging or color changing dynamics. But otherwise it totally works and
expands the usefulness of these brushes, so I don't see going beyond the
original use as that as much of an issue.