Stop CMake from orphaning binaries
On both macOS and Linux setting the VERSION target property
causes CMake to create binaries with filenames that contain the
version number in them. Every time CMake is reconfigured, that
version number changes (because it is tied to the state of Git),
and this just creates a bunch of orphaned binaries with outdated
versions in their filenames during development.

Allegedly this would set the “image version of the binary” on
Windows, but there are at least five other places where a version
number is set (four in the .rc, one in the .manifest), so we are
probably really quite good to go there without CMake needing to
do any extra thing.