Expedite announcement refreshes on urgent changes
By default, listings only get refreshed every 10 minutes or so, which is
a pretty long time to leave e.g. the title or NSFMness in a stale state.
We now refresh sessions on the next cycle when something important about
them changes, rather than waiting for them to time out.