Don't copy settings when descending into groups
This copy operation breaks when using the portable data dir for some
reason. The copy was only there because the QSettings interface is bad
and it was trying to play nice with C++ const nonsense. Now the settings
object is mutable and no copies are made. A const_cast would have worked
too, but just making it explicitly mutable does seem better than lying
about its const-ness.