Always write values with backward compatibility
Because conditionally writing the new value but unconditionally writing
the old value will lead to the old value being the one that's pulled on
next startup, which is wrong. Now the new value will be written
unconditionally too, meaning that will get loaded up next time. This
only affects very few settings, so it's not like it bloats up the
settings file or anything.